make more pages

+++ store/views.py
@@ -14,7 +14,7 @@ from .forms import *
 col = Collections()
 
 def get_cart(request):
-    # request.session['cart_item_ids'] = ''
+    request.session['cart_item_ids'] = ''
     if request.session.get('cart_item_ids'):
         cart = request.session['cart_item_ids'].split('-')[1:]
         cart_item_ids = []
@@ -22,12 +22,12 @@ def get_cart(request):
         for x in range(1, len(cart) - 1):
             real_cart.append( Cart(product_id= int(cart[x][0])  , quantity= int(cart[x][2])) )
         if request.user.is_authenticated:
-            cart_from_db_ids = [x.product_id for x in request.user.get_cart()]
+            cart_from_db_ids = [x.product_id for x in request.user.cart.all()]
             real_cart = [ x for x in real_cart if x.product_id not in cart_from_db_ids]
-            real_cart += list(request.user.get_cart())
+            real_cart += list(request.user.cart.all())
         return real_cart
     elif request.user.is_authenticated:
-        return request.user.get_cart()
+        return request.user.cart.all()
 
 
 class IndexView(ListView):
@@ -41,7 +41,7 @@ class IndexView(ListView):
         context['popular_brands'] = col.popular_brands(4)
         context['cart'] = get_cart(self.request)
         if self.request.user.is_authenticated:
-            context['wish_list'] = self.request.user.get_wish()
+            context['wish_list'] = self.request.user.wishes.all()
         return context
 
 
@@ -52,7 +52,7 @@ class CustomerCareView(IndexView):
         context = {}
         context['cart'] = get_cart(self.request)
         if self.request.user.is_authenticated:
-            context['wish_list'] = self.request.user.get_wish()
+            context['wish_list'] = self.request.user.wishes.all()
         return context
 
 
@@ -63,7 +63,7 @@ class AboutView(IndexView):
         context = {}
         context['cart'] = get_cart(self.request)
         if self.request.user.is_authenticated:
-            context['wish_list'] = self.request.user.get_wish()
+            context['wish_list'] = self.request.user.wishes.all()
         return context
 
 
@@ -75,7 +75,19 @@ class CartView(ListView):
         context = {}
         context['cart'] = get_cart(self.request)
         if self.request.user.is_authenticated:
-            context['wish_list'] = self.request.user.get_wish()
+            context['wish_list'] = self.request.user.wishes.all()
+        return context
+
+
+class CheckoutView(LoginRequiredMixin, ListView):
+    model = Cart
+    template_name = 'store/pages/checkout.html'
+    success_url = '/store/'
+
+    def get_context_data(self, **kwargs):
+        context = {}
+        context['cart'] = get_cart(self.request)
+        context['wish_list'] = self.request.user.wishes.all()
         return context
 
 
@@ -87,9 +99,31 @@ class WishListView(LoginRequiredMixin, ListView):
     def get_context_data(self, **kwargs):
         context = {}
         context['cart'] = get_cart(self.request)
-        context['wish_list'] = self.request.user.get_wish()
+        context['wish_list'] = self.request.user.wishes.all()
         return context
 
+
+class OrdersView(LoginRequiredMixin, ListView):
+    model = Order
+    template_name = 'store/pages/orders.html'
+    success_url = '/store/'
+
+    def get_context_data(self, **kwargs):
+        context = {'orders': self.request.user.orders.all()}
+        context['cart'] = get_cart(self.request)
+        context['wish_list'] = self.request.user.wishes.all()
+        return context
+
+class OrderDetailView(ListView):
+    model = Order
+    template_name = 'store/pages/order_detail.html'
+    def get_context_data(self, **kwargs):
+        order = Order.objects.get(ref=self.kwargs['ref'])
+        context = { 'order': order}
+        context['cart'] = get_cart(self.request)
+        if self.request.user.is_authenticated:
+            context['wish_list'] = self.request.user.wishes.all()
+        return context
 
 class ProductDetailView(DetailView):
     model = Product
@@ -100,74 +134,69 @@ class ProductDetailView(DetailView):
         context['related_products'] = col.related_products(product, 4)
         context['cart'] = get_cart(self.request)
         if self.request.user.is_authenticated:
-            context['wish_list'] = self.request.user.get_wish()
+            context['wish_list'] = self.request.user.wishes.all()
         return context
 
 
 class BrandDetailView(DetailView):
     model = Brand
-    template_name = 'store/pages/brand.html'
+    template_name = 'store/pages/brand_detail.html'
     def get_context_data(self, **kwargs):
         brand = Brand.objects.get(slug=self.kwargs['slug'])
-        context = { 'brand': brand}
+        context = { 'brand': brand }
+        context['popular_brand_products'] = col.popular_brand_products(brand, 6)
+        context['latest_brand_products'] = col.latest_brand_products(brand, 8)
         context['cart'] = get_cart(self.request)
         if self.request.user.is_authenticated:
-            context['wish_list'] = self.request.user.get_wish()
+            context['wish_list'] = self.request.user.wishes.all()
         return context
 
 
 class StoreView(ListView):
     model = Product
-    # get products that are 1 or more in store
-    queryset = Product.objects.filter(quantity__gte=1).order_by('-created_at')[:5]
-    context_object_name = 'products'
+    queryset = Product.objects.all()
     template_name = 'store/pages/store.html'
     success_url = '/store/'
 
     def get_context_data(self, **kwargs):
-        context = super(StoreView, self).get_context_data(**kwargs)
-        context['popular_products'] = col.popular_products(6)
+        context = {'popular_products':col.popular_products(6) }
         context['popular_brands'] = col.popular_brands(4)
         context['latest_products'] = col.latest_products(6)
         context['categories'] = col.categories(4)
         context['cart'] = get_cart(self.request)
         if self.request.user.is_authenticated:
-            context['wish_list'] = self.request.user.get_wish()
+            context['wish_list'] = self.request.user.wishes.all()
 
         return context
 
 
 class MenStoreView(ListView):
     model = Product
-    # get products that are 1 or more in store
-    queryset = Product.objects.filter(quantity__gte=1, gender='male').order_by('-created_at')[:5]
-    context_object_name = 'products'
+    queryset = Product.objects.all()
     template_name = 'store/pages/men.html'
     success_url = '/store/'
 
     def get_context_data(self, **kwargs):
-        context = super(MenStoreView, self).get_context_data(**kwargs)
+        context = { 'popular_products': col.popular_men_products(6) }
+        context['categories'] = col.categories()
         context['cart'] = get_cart(self.request)
         if self.request.user.is_authenticated:
-            context['wish_list'] = self.request.user.get_wish()
-
+            context['wish_list'] = self.request.user.wishes.all()
         return context
 
 
 class WomenStoreView(ListView):
     model = Product
-    # get products that are 1 or more in store
-    queryset = Product.objects.filter(quantity__gte=1, gender='female').order_by('-created_at')[:5]
-    context_object_name = 'products'
+    queryset = Product.objects.all()
     template_name = 'store/pages/women.html'
     success_url = '/store/'
 
     def get_context_data(self, **kwargs):
-        context = super(WomenStoreView, self).get_context_data(**kwargs)
+        context = { 'popular_products': col.popular_women_products(6) }
+        context['categories'] = col.categories()
         context['cart'] = get_cart(self.request)
         if self.request.user.is_authenticated:
-            context['wish_list'] = self.request.user.get_wish()
-
+            context['wish_list'] = self.request.user.wishes.all()
         return context
 
 
@@ -196,9 +225,95 @@ class ProfileView(LoginRequiredMixin, ListView):
     def get_context_data(self, **kwargs):
         context = {}
         context['cart'] = get_cart(self.request)
-        context['wish_list'] = self.request.user.get_wish()
+        context['wish_list'] = self.request.user.wishes.all()
         return context
 
+
+class ShippingAddressesView(LoginRequiredMixin, ListView):
+    model = ShippingAddress
+    template_name = 'store/pages/shipping_address_list.html'
+    success_url = '/store/'
+
+    def get_context_data(self, **kwargs):
+        shipping_addresses = ShippingAddress.objects.filter(user = self.request.user)
+        context = {'shipping_addresses':shipping_addresses}
+        context['cart'] = get_cart(self.request)
+        context['wish_list'] = self.request.user.wishes.all()
+        return context
+
+
+class NewShippingAddressView(LoginRequiredMixin, ListView):
+    model = ShippingAddress
+    template_name = 'store/pages/new_shipping_address.html'
+    success_url = '/store/'
+
+    def post(self, request, *args, **kwargs):
+        form = ShippingAddressForm(request.POST or None, request.user)
+        context = {'form': form}
+        context['cart'] = get_cart(self.request)
+        if form.is_valid():
+            shipping_address = ShippingAddress()
+            shipping_address.user = request.user
+            shipping= ShippingAddress.objects.filter(user=request.user, is_default=True)
+            if shipping.exists() and request.POST.get('is_default'):
+                shipping.update(is_default=False)
+                shipping_address.is_default = True
+            elif not shipping.exists():
+                shipping_address.is_default = True
+            shipping_address.zip_code = form.cleaned_data['zip_code']
+            shipping_address.address = form.cleaned_data['address']
+            shipping_address.city = form.cleaned_data['city']
+            shipping_address.state = form.cleaned_data['state']
+            shipping_address.save()
+            context['success'] = 'Your shipping address has been uploaded'
+            return render(request, 'store/pages/new_shipping_address.html', context)
+        return render(request, 'store/pages/new_shipping_address.html', context)
+
+    def get_context_data(self, **kwargs):
+        context = {}
+        context['cart'] = get_cart(self.request)
+        context['wish_list'] = self.request.user.wishes.all()
+        return context
+
+
+class ShippingAddressDetailView(LoginRequiredMixin, ListView):
+    model = ShippingAddress
+    template_name = 'store/pages/shipping_address_detail.html'
+    success_url = '/store/'
+
+    def post(self, request, *args, **kwargs):
+        form = ShippingAddressForm(request.POST or None, request.user)
+        context = {'form': form}
+        context['cart'] = get_cart(self.request)
+        if form.is_valid():
+            num = int(self.kwargs['num'])
+            shippings = ShippingAddress.objects.filter(user=request.user, is_default=True)
+            shipping_address = ShippingAddress.objects.filter(user=request.user)[num - 1]
+            if request.POST.get('is_default') and shippings.exists():
+                shippings.update(is_default=False)
+                shipping_address.is_default = True
+            shipping_address.zip_code = form.cleaned_data['zip_code']
+            shipping_address.address = form.cleaned_data['address']
+            shipping_address.city = form.cleaned_data['city']
+            shipping_address.state = form.cleaned_data['state']
+            shipping_address.save()
+            context['shipping_address'] = shipping_address
+            context['success'] = 'Your shipping address has been updated'
+            return render(request, 'store/pages/shipping_address_detail.html', context)
+        return render(request, 'store/pages/shipping_address_detail.html', context)
+
+    def get_context_data(self, **kwargs):
+        num = int(self.kwargs['num'])
+        shipping_addresses = ShippingAddress.objects.filter(user=self.request.user)
+        if shipping_addresses.count() >= num:
+            shipping_address = shipping_addresses[num - 1]
+        else:
+            shipping_address = None
+        context = {'shipping_address':shipping_address}
+        context['shipping_address_num'] = num
+        context['cart'] = get_cart(self.request)
+        context['wish_list'] = self.request.user.wishes.all()
+        return context
 
 def handle_login(request):
     if request.user.is_authenticated:
@@ -253,15 +368,15 @@ def handle_logout(request):
 @csrf_exempt
 def add_to_cart(request):
     if request.session.get('cart_item_ids'):
-        if request.POST['product_id'] + 'x' in request.session.get('cart_item_ids'):
-            product = request.POST['product_id'] + 'x'
-            cart = request.session.get('cart_item_ids')
-            index = cart.find(product)
-            new_qty = int(cart[ index + 2]) + int(request.POST['quantity'])
-            request.session['cart_item_ids'] = cart.replace(product + cart[ index + 2], \
-                product + str(new_qty))
-        else:
-            request.session['cart_item_ids'] += '-{0}x{1}'.format(request.POST['product_id'], request.POST['quantity'])
+        # if request.POST['product_id'] + 'x' in request.session.get('cart_item_ids'):
+        #     product = request.POST['product_id'] + 'x'
+        #     cart = request.session.get('cart_item_ids')
+        #     index = cart.find(product)
+        #     new_qty = int(cart[ index + 2]) + int(request.POST['quantity'])
+        #     request.session['cart_item_ids'] = cart.replace(product + cart[ index + 2], \
+        #         product + str(new_qty))
+        # else:
+        request.session['cart_item_ids'] += '-{0}x{1}'.format(request.POST['product_id'], request.POST['quantity'])
     else:
         request.session['cart_item_ids'] = '-{0}x{1}'.format(request.POST['product_id'], request.POST['quantity'])
     if request.user.is_authenticated:
@@ -280,6 +395,40 @@ def add_to_cart(request):
     return response
 
 @csrf_exempt
+def change_cart_item_qty(request):
+    if int(request.POST['new_quantity']) < 1:
+        response = JsonResponse({'status' : 'error', 'msg': 'quantity less than 0' })
+        response.status_code = 402
+        return response
+
+    if request.session.get('cart_item_ids'):
+        if request.POST['product_id'] + 'x' in request.session.get('cart_item_ids'):
+            product = request.POST['product_id'] + 'x'
+            cart = request.session.get('cart_item_ids')
+            index = cart.find(product)
+            request.session['cart_item_ids'] = cart.replace(product + cart[ index + 2], \
+                product + request.POST['new_quantity'])
+
+            response = JsonResponse({'status' : 'success', 'msg': 'quantity changed successfully' })
+            response.status_code = 200
+            return response
+        else:
+            response = JsonResponse({'status' : 'error', 'msg': 'error occured, please try again later.' })
+            response.status_code = 402
+            return response
+    if request.user.is_authenticated:
+        cart_item = Cart.objects.filter(user=request.user, product_id=request.POST['product_id']).first()
+        if cart_item:
+            cart_item.quantity = request.POST['new_quantity']
+            cart_item.save()
+            response = JsonResponse({'status' : 'success', 'msg': 'quantity changed successfully' })
+            response.status_code = 200
+            return response
+    response = JsonResponse({'status' : 'error', 'msg': 'error occured, please try again later.' })
+    response.status_code = 402
+    return response
+
+@csrf_exempt
 def remove_from_cart(request):
     if request.session.get('cart_item_ids'):
         product = '-' + request.POST['product_id'] + 'x'
@@ -290,12 +439,17 @@ def remove_from_cart(request):
         print (request.session['cart_item_ids'])
     if request.user.is_authenticated:
         cart_item = Cart.objects.filter(user = request.user, product_id = request.POST['product_id'])
-        cart_item.delete()
-        response = JsonResponse({'status' : 'success', 'msg': 'removed successfully' })
-        response.status_code = 200
-        return response
-    response = JsonResponse({'status' : 'error', 'msg': 'error occured, please try again later.' })
-    response.status_code = 402
+        if cart_item.exists():
+            cart_item.delete()
+            response = JsonResponse({'status' : 'success', 'msg': 'removed successfully' })
+            response.status_code = 200
+            return response
+        else:
+            response = JsonResponse({'status' : 'error', 'msg': 'error occured, please try again later.' })
+            response.status_code = 402
+            return response
+    response = JsonResponse({'status' : 'success', 'msg': 'removed successfully' })
+    response.status_code = 200
     return response
 
 @csrf_exempt
@@ -352,3 +506,50 @@ def empty_wish_list(request):
         response = JsonResponse({'status' : 'success', 'msg': 'wish list emptied' })
         response.status_code = 200
         return response
+
+@csrf_exempt
+@login_required(login_url='/login/')
+def make_purchase(request):
+    if not request.user.shipping_addresses.filter(is_default=True).exists():
+        response = JsonResponse({'status' : 'error', 'msg': 'your default shipping address is not set', 'shipping': True })
+        response.status_code = 402
+        return response
+    cart = get_cart(request)
+    if cart:
+        # create an order
+        order = Order(user=request.user)
+        order.shipping_address = request.user.shipping_addresses.filter(is_default=True).first()
+        order.save()
+        for item in cart:
+            # create order_items
+            order_item = OrderItem(product_id=item.product_id,
+                order=order, quantity=item.quantity, price_per_unit=item.product.price_per_unit
+            )
+            order_item.save()
+        # clear cart
+        cart.delete()
+        request.session['cart_item_ids'] = ''
+        response = JsonResponse({'status' : 'success', 'msg': 'order successfully made' })
+        response.status_code = 200
+        return response
+
+        response = JsonResponse({'status' : 'error', 'msg': 'error occured, please try again later.' })
+        response.status_code = 402
+        return response
+    response = JsonResponse({'status' : 'error', 'msg': 'your cart is empty' })
+    response.status_code = 402
+    return response
+
+@csrf_exempt
+@login_required(login_url='/login/')
+def remove_shipping_address(request):
+    shipping_address = ShippingAddress.objects.filter(user = request.user, pk = request.POST['shipping_id']).first()
+    if not shipping_address.orders.all():
+        shipping_address.delete()
+        response = JsonResponse({'status' : 'success', 'msg': 'removed successfully' })
+        response.status_code = 200
+        return response
+
+    response = JsonResponse({'status' : 'error', 'msg': 'you cannot delete this shipping_address. An order is shipping to it' })
+    response.status_code = 402
+    return response
